+++ src/ApolloProvider.tsx
@@ -3,6 +3,7 @@ import * as PropTypes from 'prop-types';
 import { Component } from 'react';
 import ApolloClient from 'apollo-client';
 import { DocumentNode } from 'graphql';
+import { ApolloContext } from './ApolloContext';
 
 import { invariant } from 'ts-invariant';
 
@@ -49,6 +50,12 @@ export default class ApolloProvider<TCache> extends Component<ApolloProviderProp
   }
 
   render() {
-    return this.props.children;
+    return ApolloContext ? (
+      <ApolloContext.Provider value={this.getChildContext()}>
+        {this.props.children}
+      </ApolloContext.Provider>
+    ) : (
+      this.props.children
+    );
   }
 }